+++ index.html
@@ -63,7 +63,6 @@
         </div>
       </div>
     </main>
-    <script src="index.js" type="module"></script>
-    <script src="modules/create_pokemon_card.js" type="module"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
