+++ index.js
@@ -4,16 +4,14 @@
  ********
  */
 
-import { constructTheCardWithAllInformation } from "./create_pokemon_card";
+import { constructTheCardWithAllInformation } from "./modules/createPokemonCard.js";
+import * as pokeAPI from "./modules/pokeAPI.js";
 
 /*
  **********
  *ELEMENTS*
  **********
  */
-
-const POKE_API_BASE_URL = "https://pokeapi.co/api";
-const VERSION = "v2";
 
 const pokedexListElement = document.getElementById("pokedex-list");
 const pokedexLoadMoreButtonElement =
@@ -41,14 +39,14 @@ let currentPokemonDetail = [];
  */
 
 function getSetOfPokemons() {
-  getPokemons().then((result) => {
+  pokeAPI.getPokemons().then((result) => {
     createNewPokemonCardsFromJSON(result);
   });
 }
 
 function loadMorePokemons() {
   addCSSAnimationClassToElement();
-  getNextSetOfPokemons(nextSetOfPokemons).then((result) =>
+  pokeAPI.getNextSetOfPokemons(nextSetOfPokemons).then((result) =>
     createNewPokemonCardsFromJSON(result)
   );
 }
@@ -63,7 +61,7 @@ async function createNewPokemonCardsFromJSON(pokemonsJSON) {
   nextSetOfPokemons = pokemonsJSON.next;
   addCSSAnimationClassToElement();
   for (const pokemon of pokemonsJSON.results) {
-    const pokemonDetails = await getPokemon(pokemon.name);
+    const pokemonDetails = await pokeAPI.getPokemon(pokemon.name);
     currentPokemons.push(pokemonDetails);
   }
   loaderElement.classList.remove("pokedex-loader-animation");
@@ -78,103 +76,20 @@ async function createNewPokemonCardsFromJSON(pokemonsJSON) {
 function createPokemonCards() {
   currentPokemons.forEach((pokemon) =>
     constructTheCardWithAllInformation(
+      pokedexListElement,
       pokemon,
       pokemon.name,
-      pokemon["sprites"]["front_default"],
-      returnTheContainersSecondType(pokemon)
+      pokemon.sprites.front_default
     )
   );
 }
 
-function createPokemonDetails(idOfPokemon, idOfHTMLElement) {
+/* function createPokemonDetails(idOfPokemon, idOfHTMLElement) {
   currentPokemonDetail = [];
-  const pokemon = getPokemon(idOfPokemon);
+  const pokemon = pokeAPI.getPokemon(idOfPokemon);
   currentPokemonDetail.push(pokemon);
   createPokemonDetailCard(idOfHTMLElement);
-}
-
-/*
- *****
- *API*
- *****
- */
-
-/**
- * Fetches the given URL with an fetch method
- * @param {string} URL
- * @param {object} options
- * @returns the response in JSON format of the fetched URL
- */
-async function fetchAPI(URL, options) {
-  const response = await fetch(URL, options);
-  if (response.ok) {
-    return response.json();
-  } else {
-    alert("ERROR");
-  }
-}
-
-async function fetchHTMLFile(file, options) {
-  const response = await fetch(file, options);
-  if (response.ok) {
-    return response.text();
-  } else {
-    alert("ERROR");
-  }
-}
-
-/**
- * Calls the function fetchAPI() with the pokemonURL constant and with the method "GET"
- * @returns The response in JSON of the pokemonURL
- */
-function getPokemons() {
-  return fetchAPI(`${POKE_API_BASE_URL}/${VERSION}/pokemon`, { method: "GET" });
-}
-
-/**
- * Calls the function fetchAPI() with the nextURL parameter and with the method "GET"
- * The nextURL parameter refers to the url given by the api pointing to the next of comming pokemons
- * @returns The response in JSON of the nextURL
- */
-function getNextSetOfPokemons(nextURL) {
-  return fetchAPI(nextURL, { method: "GET" });
-}
-
-function getPokemon(nameOrIdOfPokemon) {
-  return fetchAPI(
-    `${POKE_API_BASE_URL}/${VERSION}/pokemon/${nameOrIdOfPokemon}`,
-    { method: "GET" }
-  );
-}
-
-function getAbilities(nameOfPokemon) {
-  return fetchAPI(`${POKE_API_BASE_URL}/${VERSION}/ability/${nameOfPokemon}`, {
-    method: "GET",
-  });
-}
-function getCharacteristic(nameOfPokemon) {
-  return fetchAPI(
-    `${POKE_API_BASE_URL}/${VERSION}/characteristic/${nameOfPokemon}`,
-    { method: "GET" }
-  );
-}
-
-function getPokemonSpecies(nameOfPokemon) {
-  return fetchAPI(
-    `${POKE_API_BASE_URL}/${VERSION}/pokemon-species/${nameOfPokemon}`,
-    { method: "GET" }
-  );
-}
-
-function returnTheContainersSecondType(result) {
-  return result.types[1] === undefined
-    ? ""
-    : '<div class="pokedex-type-word-container" id="second-pokemon-type-' +
-        result.id +
-        '"><p>' +
-        result.types[1].type.name +
-        "</p></div>";
-}
+} */
 
 /*
  *****************
@@ -186,7 +101,7 @@ async function searchForAPokemon() {
   pokedexListElement.innerHTML = "";
   currentPokemons = [];
   addCSSAnimationClassToElement();
-  const pokemonDetails = await getPokemon(searchElement.value.toLowerCase());
+  const pokemonDetails = await pokeAPI.getPokemon(searchElement.value.toLowerCase());
 
   currentPokemons.push(pokemonDetails);
   loaderElement.classList.remove("pokedex-loader-animation");
@@ -203,7 +118,6 @@ async function searchForAPokemon() {
  *************************************
  */
 
-
 // For details of the pokemon
 function constructDetailCard(
   idOfElement,
@@ -219,60 +133,6 @@ function constructDetailCard(
     "pokemon-type-",
     pokemonDetails.id
   );
-}
-
-function applySecondPokemonTypeIfExists(pokemonDetails) {
-  if (pokemonDetails.types[1] != undefined) {
-    applyPokemonTypes(
-      pokemonDetails.types[1].type.name,
-      "second-pokemon-type-",
-      pokemonDetails.id
-    );
-  }
-}
-
-function applyTheBackgroundColor(nameOfType, whichPokemonId, pokemonId) {
-  let lastElementFromArray = document.getElementById(
-    whichPokemonId + pokemonId
-  );
-  lastElementFromArray.classList.add(nameOfType);
-}
-
-function emojiSetting(emoji, pokemonId) {
-  let lastElementFromArray = document.getElementById(
-    "emoji-pokemon-" + pokemonId
-  );
-  lastElementFromArray.innerHTML = emoji;
-}
-
-function applyPokemonTypes(nameOfFirstType, whichPokemonId, pokemonId) {
-  const type_mapping = {
-    grass: { color: "background-grass-type", emoji: "🌿" },
-    water: { color: "background-water-type", emoji: "💧" },
-    fire: { color: "background-fire-type", emoji: "🔥" },
-    normal: { color: "background-normal-type", emoji: "⭐" },
-    poison: { color: "background-poison-type", emoji: "☠️" },
-    flying: { color: "background-flying-type", emoji: "🪶" },
-    bug: { color: "background-bug-type", emoji: "🐞" },
-    fairy: { color: "background-fairy-type", emoji: "🧚‍♀️" },
-    psychic: { color: "background-psychic-type", emoji: "👁️‍🗨️" },
-    electric: { color: "background-electric-type", emoji: "⚡" },
-    ground: { color: "background-ground-type", emoji: "🗿" },
-    fighting: { color: "background-fighting-type", emoji: "🥊" },
-    dragon: { color: "background-dragon-type", emoji: "🐉" },
-    rock: { color: "background-rock-type", emoji: "🪨" },
-    ice: { color: "background-ice-type", emoji: "❄️" },
-    ghost: { color: "background-ghost-type", emoji: "👻" },
-    steel: { color: "background-steel-type", emoji: "🛡️" },
-    dark: { color: "background-dark-type", emoji: "🌙" },
-  };
-
-  applyTheBackgroundColor(
-    type_mapping[nameOfFirstType].color,
-    whichPokemonId,
-    pokemonId
-  );
-  emojiSetting(type_mapping[nameOfFirstType].emoji, pokemonId);
 }
 
 /*
