+++ modules/stylePokemonCard.js
@@ -1,0 +1,54 @@
+export function applySecondPokemonTypeIfExists(pokemonDetails) {
+  if (pokemonDetails.types[1] != undefined) {
+    applyPokemonTypes(
+      pokemonDetails.types[1].type.name,
+      "second-pokemon-type-",
+      pokemonDetails.id
+    );
+  }
+}
+
+export function applyTheBackgroundColor(nameOfType, whichPokemonId, pokemonId) {
+  let lastElementFromArray = document.getElementById(
+    whichPokemonId + pokemonId
+  );
+  lastElementFromArray.classList.add(nameOfType);
+}
+
+export function emojiSetting(emoji, pokemonId) {
+  let lastElementFromArray = document.getElementById(
+    "emoji-pokemon-" + pokemonId
+  );
+  lastElementFromArray.innerHTML = emoji;
+}
+
+export function applyPokemonTypes(nameOfFirstType, whichPokemonId, pokemonId) {
+  const type_mapping = {
+    grass: { color: "background-grass-type", emoji: "🌿" },
+    water: { color: "background-water-type", emoji: "💧" },
+    fire: { color: "background-fire-type", emoji: "🔥" },
+    normal: { color: "background-normal-type", emoji: "⭐" },
+    poison: { color: "background-poison-type", emoji: "☠️" },
+    flying: { color: "background-flying-type", emoji: "🪶" },
+    bug: { color: "background-bug-type", emoji: "🐞" },
+    fairy: { color: "background-fairy-type", emoji: "🧚‍♀️" },
+    psychic: { color: "background-psychic-type", emoji: "👁️‍🗨️" },
+    electric: { color: "background-electric-type", emoji: "⚡" },
+    ground: { color: "background-ground-type", emoji: "🗿" },
+    fighting: { color: "background-fighting-type", emoji: "🥊" },
+    dragon: { color: "background-dragon-type", emoji: "🐉" },
+    rock: { color: "background-rock-type", emoji: "🪨" },
+    ice: { color: "background-ice-type", emoji: "❄️" },
+    ghost: { color: "background-ghost-type", emoji: "👻" },
+    steel: { color: "background-steel-type", emoji: "🛡️" },
+    dark: { color: "background-dark-type", emoji: "🌙" },
+  };
+
+  applyTheBackgroundColor(
+    type_mapping[nameOfFirstType].color,
+    whichPokemonId,
+    pokemonId
+  );
+  emojiSetting(type_mapping[nameOfFirstType].emoji, pokemonId);
+}
+
